refactor(core): declare the extension root manifest once
